x86/tdx: Disable kvmclock in TDX
TDX guests have a special secured TSC enforced by the TDX module to
provide secure time to guests. This only works when only RDTSC is used
to derive the time. So disable kvmclock in TDX because it uses host
information too, which could be abused by a malicious hypervisor.

Define a new SECURE_TIME prot flag for this.

Note there are other time sources of course, but kvmclock has higher
priority than TSC, so disabling it is most important.

Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: Kuppuswamy Sathyanarayanan <sathyanarayanan.kuppuswamy@linux.intel.com>
[ Zhiquan Li: amend commit log and resolve the conflict ]
Signed-off-by: Zhiquan Li <zhiquan1.li@intel.com>