KVM: Introduce API to get the Target VM's kvm struct using PID or vmfd
Introduces a new API, kvm_get_target_kvm(), to find the target VM's kvm
struct using a caller-provided target_vm_id. The target_vm_id parameter
allows callers to specify either the process ID (PID) or the vmfd (which
is returned by KVM_CREATE_VM) of the target VM. If the target VM is
found, its kvm struct is returned to the caller, otherwise, NULL is
returned.

This function is particularly useful when a KVM implementation for a guest
needs to access the KVM struct of another guest. For instance, an Intel TD
guest may need to bind to a special service TD guest, which necessitates
access to certain fields stored in the target KVM struct.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>