KVM: TDX: Bypass tdh_mem_range_block() when TD hasn't been initialized
The TDH.MEM.RANGE.BLOCK seamcall requires that the TD has been
initialized. In the live migration case, the destination TD is
initialized after the TD states are imported, though its hkid has been
assigned and configured on TD launch.

tdx_sept_zap_private_spte() can be called for private-to-shared page
conversion. The function currently zaps the private spte before removing
it (see handle_removed_private_spte()). Such page conversion can occur
during live migration and in the case of destination TD, TD is not
running yet (and not done initialization), so bypass the "zap private
spte" step by returning 0. This allows the function to proceed to the
next step (i.e. remove the spte of the pages that have been imported
to the destination TD).

Signed-off-by: Wei Wang <wei.w.wang@intel.com>