KVM: TDX: Add TDG.VP.VMCALL.SERVICE
TDG.VP.VMCALL.SERVICE is a sub function of TDG.VP.VMCALL used by a service
TD to communicate with KVM (namely, to perform a service call) . The
tdvmcall_service struct follows the data format defined in the GHCI spec
for the communication.

The service call is made by the service TD with two buffers (the data
inside follows the tdvmcall_service struct format) passed to KVM:
- A 4KB aligned command buffer. The data inside is a command filled by
  the service TD, and will be read by KVM;
- A 4KB aligned response buffer. KVM will fill the response data after
  handling the command, and the respnse data will be read by the service
  TD.

The current KVM implementation doesn't support asynchronously handling the
service call, i.e., sending interrupts with the vector specified in the
service call to the service TD. So the service TD is expected to poll on
responses. If the response isn't ready, the vcpu that made the service
call will be halted by KVM, and be waken up later when response is ready
to read.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>