KVM: TDX: Add seamcalls TDH.EXPORT.BLOCKW and TDH.EXPORT.UNBLOCKW
TDH.EXPORT.BLOCKW is called by KVM to block a list of TD private pages
for writing (i.e. write-protection). The GPA list format (defined in the
TDX ABI spec) is used for the call, and the physical address of the list
is contained in the @gpa_list_info parameter. If a page failed to be
blocked for writing, the error is marked in the related entry of the GPA
list page, but the seamcall doesn't abort, and the number of pages that
failed will be includes in the returned status (i.e. in rax).

TDH_EXPORT_BLOCKW is interruptible. If a pending interrupt is detected
during operation, it returns TDX_INTERRUPTED_RESUMABLE with rcx holding
the next entry in the GPA list page to continue. The caller can re-invoke
the seamcall immediately after handling the interrupt.

TDH.EXPORT.UNBLOCKW is called by KVM to remove the write-blocking of a
single 4KB TD private page which was previously write-blocked. This is
usually called when a guest tries to write to its pages which has been
write-blocked during live migration. The GPA of the page that needs to
be unblocked for writing is contained in the @ept_info parameter (its
format is defined in the TDX ABI spec).

Signed-off-by: Wei Wang <wei.w.wang@intel.com>