KVM: TDX_MIG: Pre-allocate SEPT tables after immutable states imported
In live migration, pre-allocate the secure EPT tables (except the leaf
entries) of the destination TD after the immutable states are imported.
This facilitates later in-place memory import. The SEPT leaf entries are
added when importing the private page with in-place import, where the
pages imported by TDH.IMPORT.MEM become the SEPT leaf pages. This requires
the SEPT tables to be pre-allocated.

The SEPT tables can only be pre-allocated after the immutable state gets
imported, as the importing the immutable states finishes a prelimilary
initialization (e.g. the level of SEPTs) for the destination TD.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>