KVM: TDX: Add td_initialized check in tdx_init_mem_region()
Add a check for td_initialized in the tdx_init_mem_region() function. If a
KVM_INIT_MEM_REGION request is received to initialize a memory region
before the TD is initialized, the check will report errors to userspace.
This enhancement aids in alerting the userspace implementation about
potential bugs or necessary modifications.

A practical example of its utility is during the launch of a destination
TD in a live migration scenario. In such a case, the KVM_INIT_MEM_REGION
command should be bypassed. This is because the memory region setup (e.g.
for pages with TDVF binary loaded) via KVM_INIT_MEM_REGION is not needed
for the destination TD, as the memory will be migrated from the source TD.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>