KVM: TDX: Handle TDG.VP.VMCALL.SERVICE.QUERY
The QUERY service (defined in the GHCI spec) is one of the services
supported by TDG.VP.VMCALL.SERVICE. It's used by a service TD to query if
any of the services (identified by a 128-bit service GUID) that can be
requested via TDG.VP.VMCALL.SERVICE.* are supported by KVM. The 128-bit
GUIDs for various services defined in the spec are translated into enum
tdvmcall_service_id type for easy referencing.

If the queried service is supported, TDVMCALL_SERVICE_QUERY_S_SUPPORTED
is returned in the status field. Otherwise, the status field is set to
TDVMCALL_SERVICE_QUERY_S_UNSUPPORTED.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>