KVM: x86/cgm: Add KVM_CGM_START ioctl
KVM_CGM_START is used by userspace to request the secure firmware (e.g.
TDX module or SVSM) via KVM to start (i.e. kick off) a migration session.
Depending on the secure firmware design, it may require to migrate some
initialization data from the source to destination to do a preliminary
initialization (e.g. TDX's immutable state export and import).
"struct kvm_cgm_data" is added for userspace to send a buffer to KVM,
and the buffer is passed to the vendor's callback to load the
initialization data from the secure firmware on the source or set the
initialization data to the secure firmware on the destination. On the
source side, the callback is responsible for updating the @size field
with the size of initialization data that has been loaded into buffer
so that userspace can know the amount of data to send to the destination.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>