KVM: x86/cgm: Add KVM_CGM_GET_EPOCH_TOKEN ioctl
KVM_CGM_GET_EPOCH_TOKEN is used by userspace, on the live migration source
side, to request the secure firmware (e.g. TDX module) via KVM to get an epoch
token. The token acts as a boundary marker between migration epochs (also known
as pre-copy iterations). This mechanism helps prevent replay attacks.

The epoch token data is utilized by the vendor-specific secure firmware. VMM
does not need to parse the token data. Instead, the VMM's role is to transfer
the token data from the source to the destination secure firmware. Userspace
provides a buffer via "struct kvm_cgm_data" to the secure firmware via KVM, and
the buffer will be loaded with the token data on return of the ioctl.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>