KVM: TDX_MIG: Add tdx_mig_get_epoch_token()
The tdx_mig_get_epoch_token() callback function is invoked via the
.get_epoch_token() hook to get an epoch token from the TDX module via
the TDH.EXPORT.TRACK seamcall.

tdx_mig_stream_export_track() is made as a wrapper that allows callers
to export either an epoch token or a start token, as determined by the
@need_start_token parameter. In the case of tdx_mig_get_epoch_token(),
which is called via the KVM_CGM_GET_EPOCH_TOKEN ioctl, @need_start_token
is set to false to export an epoch token from the TDX module.

The buffer for getting the token data is supplied by the userspace, so it
is pinned prior to sharing with the TDX module and unpinned once the data
export is complete. This prevents the page from being swapped out while in
use by the TDX module.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>