KVM: TDX_MIG: Add tdx_mig_get_vcpu_state()
The tdx_mig_get_vcpu_state() callback is invoked via the .get_vcpu_state()
hook to get a TD's vcpu state from the TDX module. There are some TDX
specific states and operations that need to be handled during the TD
migration downtime period. These states are not universally applicable to
the generic live migration flow, hence they are not incorporated via new
hooks to the CGM abstraction layer. Instead, they are managed via the
.get_vcpu_state() hook in conjunction with the vcpu states. This approach
of handling TDX-specific migration states/operations within the tdx_mig
driver simplifies the abstraction layer, which is designed to be
vendor-neutral.

When the request to get the first vcpu state is made, additional
operations are executed along with exporting the vcpu state:
1) The TDX-enforced downtime period is initiated via the TDH.EXPORT.PAUSE
   seamcall.
2) The TD-scope mutable states are exported via the TDH.EXPORT.STATE.TD
   seamcall into the buffer provided by @data->uaddr.
3) The vcpu states are exported via the TDH.EXPORT.STATE.VP seamcall into
   the buffer provided by @data->uaddr.

From the perspective of the CGM abstraction or userspace VMM, it appears
as a data blob that contains the TD's vcpu state. However, at the tdx_mig
driver layer, the blob data includes an additional TD-scope mutable
states. The size of the blob data (i.e., the kvm_cgm_data->size field) is
the cumulative size of the two states. To assist the destination tdx_mig
driver in parsing the blob of two states, "struct tdx_mig_mbmd_ext" (an
extension to the TDX's MBMD data per state bundle used by the tdx_mig
drivers) introduces a new field, "nr_pages", to record the size of each
state bundle.

Once the last vcpu state is exported, TDH.EXPORT.TACK is invoked to export
the "start token", which is necessary for the TDX module to permit the
destination TD to resume and run.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>