KVM: X86: Introduce kvm_get_supported_cpuid_internal()
TDX module reports a set of configurable CPUIDs. Directly report these
bits to userspace and allow them to be set is not good nor right.
If a bit is unknown/unsupported to KVM, it should be reported as unsupported
thus inconfigurable to userspace.

Introduce and export kvm_get_supported_cpuid_internal() for TDX to get
the supported CPUID list of KVM. So that TDX can use it to cap the
configurable CPUID list reported by TDX module.

Signed-off-by: Xiaoyao Li <xiaoyao.li@intel.com>