KVM: TDX_MIG: Add tdx_mig_start()
The tdx_mig_start() callback is invoked via the .cgm_start() hook to
perform TDX specific migration start process:
1) Wait for migration prepare to complete;
2) Kick off a migration session via export immutable states on the source
   or import immutable states on the destination.

To achieve 1), the atomic flag, "migration_prepare_done", is set when
MigTD reports that the preparation is done successfully. tdx_mig_start()
polls the flag and tentatively sleep for 10ms if not set per check.
Return an error to userspace if the waiting time exceeds 10 seconds, as
the preparation might have encountered an error somewhere.

For 2), the buffer provided by userspace is pinned before sharing to TDX
module to export/import the immutable state on source/destination and
unpinned after that's done. This prevents the page from being swapped out
while the TDX module is using them. There is an extra piece of metadata
(wrapped via "struct tdx_mig_mbmd_ext") appended to the end of the MBMD
data in the MBMD buffer. The metadata is used for communication between
the tdx_mig drivers on the source and destination. A driver version number
is included in the metadata for a compatibility check between the tdx_mig
drivers on the source and destintion, as there will be driver layer
communications to prepare for data export/import. So the driver version
check is put at the start of a migration session. If the check fails, just
stops the migration process.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>