KVM: TDX: Allocate a gpa list to be used by TDH.EXPORT.BLOCKW
Dirty page logging via TDH.EXPORT.BLOCKW is performed on the source TD in
live migration, and it requires to input a list of GPAs which need to be
blocked for writing. TDH.EXPORT.BLOCKW doesn't share the gpa list that is
allocated per migration stream for TDH.EXPORT.MEM to use, because
1) TDH.EXPORT.BLOCKW could be called while TDH.EXPORT.MEM is in progress;
2) The list size is 4KB, so having two separate list doesn't consume too
   much memory while simplifying the design.

The gpa list is allocated when migration prepare is requested and
destroyed when the TD is shut down. It could be further optimized with
destroying the list when a migration session ends.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>