KVM: TDX_MIG: Add tdx_restore_private_page()
The tdx_restore_private_page() callback invokes the TDH.EXPORT.RESTORE
seamcall to restore the state of a TD private page to its original
non-exported state (maintained inside the TDX module). This is required
by TDX module on the source platform in live migration to allow the TD
private pages to be re-exported in a new migration session later.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>