KVM: TDX: Retry seamcall on recoverable errors
Errors returned by the TDX module with the non-recoverable bit (62)
cleared hints that it is likely to be a recoverable error. For
example, TDX_OPERAND_BUSY may be returned when the seamcall fails
to acquire a TDX internal lock. KVM needs to retry the seamcall in
those recoverable error cases. The seamcall shouldn't be retried for
unlimited number of times in case anything logically goes wrong, so set
a relatively large limit, TDX_SEAMCALL_RETRY_MAX, to capture any
unexpected errors. The TDX_SEAMCALL_RETRY_MAX is defined to be larger
than RDRAND_RETRY_LOOPS (i.e., 10), as in some cases (e.g. live migration
related seamcalls), the TDX internal lock may be hold longer and require
other the seamcall callers waiting for the lock to re-try more than 10
times.

Update tdx_seamcall to retry on the recoverable errors. Some recoverable
errors require some conditions to be corrected before retrying, so
return to the callers to take some actions when encountering those
errors:
- TDX_VCPU_ASSOCIATED: indicates that the vcpu has been associated to
  another pcpu. Caller needs to disassociate the vcpu before retry.
- TDX_VCPU_NOT_ASSOCIATED: indicates that the vcpu hasn't been
  associated to any pcpu. This is used with tdh_vp_flush, and it
  doesn't need to retry to do the flush.
- TDX_INTERRUPTED_RESUMABLE: caller usually needs to set a "resume"
  flag before retry the seamcall.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>