KVM: TDX: Add TDH.SERVTD.BIND and TDH.SERVTD.PREBIND seamcalls
A TD can be bound to another TD via the TDH.SERVTD.BIND seamcall. This
is termed "binding" in the TDX architecture spec. Binging is usually
performed between a user TD (i.e., a TD owned by an end user) and a
service TD (a TD usually deployed by CSPs to provide services, e.g. assist
in live migration of the user TD by setting up migration session keys).
After binding is done, the service TD will be able to access to control
structure (e.g. TDCS) fields of a user TD using TDCALLs, so the service
TD is made part of the user TD's TCB via being included in the user TD's
measurement, which is finalized by the TDH.MR.FINALIZE seamcall invoked
during the user TD launch flow.

A successful invocation of TDH.SERVTD.BIND seamcall requires that the
service TD has been running as the TDX module needs to find the related
info from the running service TD to calculate a service TD hash (i.e.
SERVTD_INFO_HASH described in the TDX Architecture spec), and the hash
is included in the user TD's measurement.

The TDH.SERVTD.BIND seamcall can be called either 1) during the user TD
launch flow (i.e. before TDH.MR.FINALIZE) or 2) when the user TD has been
running. For 2), pre-binding via TDH.SERVTD.PREBIND is required to be
called before TDH.MR.FINALIZE. In this case, SERVTD_INFO_HASH is provided
via the TDH.SERVTD.PREBIND seamcall so that it can be included in the
user TD's measurement finalized via TDH.MR.FINALIZE.

The main purpose of pre-binding is to remove the requirement of having
the service TD be running when the user TD is launched. This facilitates
the build of a transient service TD. For example, a service TD to assist
in live migration can be launched on demand when live migration is
initiated, and after the service is provided, the service TD can be
terminated.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>