KVM: TDX: Handle TDG.VP.VMCALL.SERVICE.MIGTD and its WaitForRequest cmd
Add support for the MigTD service (defined in the GHCI spec), which is
requested by a service TD via TDG.VP.VMCALL.SERVICE. The MigTD service
supports a bunch of commands, which are typically used by MigTD (i.e., a
service TD) to communiacte with KVM to preparare for live migration of a
user TD (which has been bound to MigTD via TDH.SERVTD.BIND). The service
specific data format, struct tdvmcall_service_migtd, is defined based on
the defintion in GHCI.

WaitForRequest (defined as TDVMCALL_SERVICE_MIGTD_CMD_WAIT) is one of the
commands (under the MigTD service) used by MigTD to check from KVM if
there is an migration preparation operation that needs to perform. The
command specific data, struct migtd_basic_info, is also defined based on
the defination in GHCI (named MIGTD_MIGRATION_INFORMATION_HOB_DATA there).
It contains the basic info that's required by MigTD to perform migration
preparation, for example, the binding handle and uuid of the target user
TD (returned from TDH.SERVTD.BIND) will be used by MigTD's TDG.SERVTD.WR
TDCALL to set a migration session key for the user TD. The other two pieces
of info, socket info and policy info, defined in the GHCI spec are not
supported currently, and can be added later when needed.

If there is no migration preparation needed at the time of the check via
the WaitForRequest command, the MigTD's vCPU that makes the service
call will be halted and resumed later when a preparation request is
ready.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>