KVM: Add kvm_pre_fault_private_memory_nonleaf_all()
The kvm_pre_fault_private_memory_nonleaf_all() function allocates the
stage-2 page tables (e.g. secure EPT in the TDX case) for all the private
memory pages of a guest. The TDX migration driver is a user of this
function, which allocates the secure EPT tables for the destination TD
when migration starts (i.e., after the immutable TD state is imported),
so that later private memory pages can be imported as leaf pages into
the TD's secure EPT tables.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>