KVM: x86/tdp_mmu: return flush if no need to zap the leafs
The tdp_mmu_zap_leafs() function is a generic function not only for the
private memory but also for the shared memory, the combination which
doesn't make sense to private memory mappings doesn't mean the flush
flag can be changed for the shared memory mappings, especially the
function is used by traverse a range.

So don't change the value of "@flush" for the case that requesting
zaping a shared memory mapping but encounting a private page, just keep
the original value.

The issue and fix was reported in link.

Link: https://lore.kernel.org/all/Zh8yHEiOKyvZO+QR@chao-email/
Signed-off-by: Zhiquan Li <zhiquan1.li@intel.com>