KVM: TDX_MIG: Prebind with a MigTD instance
Pre-bind the user TD with a MigTD instance during the TD initialization
time. Add a 48-bit field, migtd_hash[6], to kvm_tdx_init_vm uAPI for
userspace to send the SERVTD_INFO_HASH of the MigTD instance, which will
be passed as a parameter to the TDH.SERVTD.PREBIND seamcall. This allows
the MigTD instance to be bound later when it's launched, meaning that
the MigTD instance doesn't have to be running at the user TD launch
time.

The tdx_mig.c file is added to have all the tdx migration related
functions put here. This is more readable than having them mixed with
other funcations in tdx.c.

TODO: make TDX_MIGTD_ATTR configurable, e.g., as a kvm-intel's param.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>