KVM: TDX: Add the seamcall TDH.MIG.STREAM.CREATE
The TDH_MIG_STREAM_CREATE seamcall creates a migration stream and its
migsc control structure in the TDX module. Migration stream is a TDX
concept and it is analogous to a "channel" in the TDX module that could
be used by an operating system level thread (i.e. a migration thread) for
TD states (memory, vcpu etc.) export and import. Multiple TDX streams can
be created and assigned to multiple threads to achieve multi-threaded
data export/import from the TDX module.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>