KVM: x86/mmu: Support tdp_mmu for write-blocking private pages
The .write_block_private_pages() hook is added for tdp_mmu to call into
the vendor-specific callbacks to request the vendor's secure firmware for
write-blocking pages. It returns 0 when the write-blocking operation is
done successfully.

Currently, only one page is write-blocked at a time by tdp_mmu, but the
.write_block_private_pages() interface for vendor-specific callbacks is
designed to support passing an array of GFNs (Guest Frame Number) for
batch write-blocking. The batch support can be extended for tdp_mmu later.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>