KVM: TDX_MIG: Add tdx_mig_set_vcpu_state()
The tdx_mig_set_vcpu_state() callback is invoked via the .set_vcpu_state()
hook to set a TD's vcpu state to the TDX module. Certain TDX-specific
states, which need to be handled during the TD migration downtime period,
are not universally applicable to the generic live migration flow, so they
are packaged together with the vcpu state, instead of using new hooks
with TDX-specific semantics added to the CGM abstraction layer, which
intends to be vendor-neutral.

Upon the request to set the first vcpu state, the TD-scope mutable state
bundle is initially extracted from the blob data and imported into the TDX
module via the TDH.IMPORT.STATE.TD seamcall. This step is mandated by the
TDX migration architecture to be executed prior to importing any vcpu
state.

The tdx_td_post_init() function completes the KVM side TD initialization,
such as initializing TD-scope values by reading from the TDX module.
Therefore, it is invoked after the TD-scope states are imported.
Similarly, tdx_td_vcpu_post_init() is called to finalize the vcpu
initialization (e.g., configuring the vcpu-scope fields) after the vcpu
state is imported.

Once the last vcpu state is imported, the TDH.IMPORT.TACK seamcall is
invoked to import the "start token". This token is essential for the TDX
module to allow the destination TD to resume running.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>