KVM: TDX_MIG: Add tdx_mig_end()
The tdx_mig_end() callback is invoked via the .cgm_end() hook to end the
TDX migration. The @abort parameter signifies whether the termination is
due to an abort in migration. If it is a migration abort, a call to
kvm_mmu_restore_private_pages() is made to restore the states of the
private pages that have been exported. This is required by the TDX
module to ensure those TD private pages to be re-migratable.

In the scenario where the migration is not aborted (i.e., the migration
process is completed successfully), the destination TD is then mandated by
the TDX module to invoke the TDH,IMPORT.END seamcall. This transitions the
TD into a runnable state. For the source TD, no operations are required
and it will be destroyed.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>