KVM: TDX_MIG: Add tdx_mig_prepare()
The tdx_mig_prepare() function is invoked via the .cgm_prepare hook to
perform TDX specific operations to support migration preparation for a
TD guest, which include binding the user TD to a MigTD instance (i.e. a
service TD) and notifying the MigTD instance to start the preparation
work.

The tdx_binding_info struct is created to manage the data that is required
by MigTD to do the preparation work (details documented in the GHCI for
Intel TDX spec). The binding_info memory is allocated by the user TD and
shared with MigTD to read the data, so the users_count is initialized to
2. The count will be decremented by 1 when user TD or MigTD shuts down,
when reaching zero (i.e. both MigTD or user TD terminates), the memory
will be freed.

The tdx_mig.c file is added to have all the tdx migration related
functions put here. This is more readable than having them mixed with
other funcations in tdx.c.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>