KVM: TDX_MIG: Add tdx_mig_enable_cap()
The tdx_mig_enable_cap() callback is invoked via the .cgm_enable_cap()
hook to negotiate the caps with userspace. There are three types of state
bundles that require to use multiple pages, i.e., the immutable states,
the memory state bundle, and the vcpu state bundle. The calulation
compares the three cases and return the maximum one to userspace (via the
nr_ubuf_pages field) to ensure that the userspace allocated buffer is able
to cover all of the bundle data. Currently, only 1 migration stream is
supported, so 1 is set to the nr_threads field.

tdx_mig_state is a per-TD scope struct to manage its migration related
info. The above caps (i.e. nr_streams and nr_ubuf_pages) are recorded in
the struct for later uses during the migration session (e.g. create the
number of migration streams).

Signed-off-by: Wei Wang <wei.w.wang@intel.com>