KVM: TDX_MIG: Add "struct tdx_mig_stream" to manage per-stream info
"Migration stream" is a TDX concept (analogous to a "channel") for VMM
to export private memory and non-memory states of a TD guest during live
migration. The stream is used with an MBMD (Migration Bundle MetaData)
buffer and various lists of buffers by the memory (e.g. TDH.EXPORT.MEM)
and non-memory state (e.g.TDH.EXPORT.STAT.VP) seamcalls.

Add "struct tdx_mig_stream" to manage those lists. The lists of buffers
are per-stream, meaning that later multi-stream extensions will have
multiple "struct tdx_mig_stream"s, and tdx_mig_stream_buffer_setup() will
be called for each stream to allocate the lists for each stream, so that
data parallelism can be achieved by streams' own buffers.

Note: only the lists are allocated here. Buffers will be added to the
lists when VMMs sends them to export/import data later. The lists are
allocated when migration prepare is requested and freed when the TD is
destroyed. This will be further optimized to have the lists freed at the
time when a migration session ends.

In addition to the lists, the "struct pages" array is added for use with
pin_user_pages() on the ubufs (i.e., userspace buffers). The buffers are
allocated per-stream to achieve parallel export/import of the guest
private memory data.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>