KVM: x86/cgm: Add ioctl KVM_CGM_END
KVM_CGM_END is used by userspace, either on the source or destination side
in live migration, to request the secure firmware (e.g. TDX module) via
KVM to end the migration session. The "abort" parameter is employed to
denote whether the session termination is due to a migration abort (1) or
not (0). In the event of a session abort, the secure firmware on the
source side may perform certain steps (e.g., in the TDX case, restore the
states of the private pages that have been migrated) to ensure that the TD
is re-migratable later.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>