KVM: TDP_MMU: put the folio when a private page is skipped during import
kvm_gmem_get_pfn() increases the reference count of the folio (i.e., the
head/first page of a hugepage compound). The increase is unnecessary when
a private page is not successfully imported (indicated by the skip flag
set by the vendor-specific implementation). So the reference count should
be decremented in this case. Otherwise, some pages on the destination side
may not be released due to the unexpectedly increased reference count,
which locks the entire 2MB hugepages when the destination TD terminates.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>